+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## 1.0.0-beta.18
+### Features
+* add R7RS `guard` macro
 ### Bugfix
 * remove evaluating of async list data as first argument